sink: rename mq producer package (#877)
* rename_mq_producer_package

* refine package name
