+++ cdc/sink/producer/kafka/kafka.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mqProducer
+package kafka
 
 import (
 	"context"
@@ -30,8 +30,8 @@ import (
 	"go.uber.org/zap"
 )
 
-// KafkaConfig stores the Kafka configuration
-type KafkaConfig struct {
+// Config stores the Kafka configuration
+type Config struct {
 	PartitionNum      int32
 	ReplicationFactor int16
 
@@ -44,8 +44,8 @@ type KafkaConfig struct {
 }
 
 // NewKafkaConfig returns a default Kafka configuration
-func NewKafkaConfig() KafkaConfig {
-	return KafkaConfig{
+func NewKafkaConfig() Config {
+	return Config{
 		Version:           "2.4.0",
 		MaxMessageBytes:   512 * 1024 * 1024, // 512M
 		ReplicationFactor: 1,
@@ -192,7 +192,7 @@ func (k *kafkaSaramaProducer) run(ctx context.Context) error {
 }
 
 // NewKafkaSaramaProducer creates a kafka sarama producer
-func NewKafkaSaramaProducer(ctx context.Context, address string, topic string, config KafkaConfig, errCh chan error) (*kafkaSaramaProducer, error) {
+func NewKafkaSaramaProducer(ctx context.Context, address string, topic string, config Config, errCh chan error) (*kafkaSaramaProducer, error) {
 	log.Info("Starting kafka sarama producer ...", zap.Reflect("config", config))
 	cfg, err := newSaramaConfig(ctx, config)
 	if err != nil {
@@ -298,7 +298,7 @@ func kafkaClientID(role, captureAddr, changefeedID, configuredClientID string) (
 }
 
 // NewSaramaConfig return the default config and set the according version and metrics
-func newSaramaConfig(ctx context.Context, c KafkaConfig) (*sarama.Config, error) {
+func newSaramaConfig(ctx context.Context, c Config) (*sarama.Config, error) {
 	config := sarama.NewConfig()
 
 	version, err := sarama.ParseKafkaVersion(c.Version)
